requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
dnspython==1.16.0
frozenlist==1.4.1
idna==3.7
loguru==0.7.2
magic-filter==1.0.12
motor==3.4.0
multidict==6.0.5
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
python-dotenv==1.0.1
typing_extensions==4.11.0
yarl==1.9.4
